restructure note files
