+++ frontend/assets/NoteIcon.tsx
